Extracts WithCurrentTime HOC into hook (TODO: specs)
